(interpreter) Extract TypeValidator inner classes
